[SofaKernel] Move SofaException from helper to core
Because exception has access to the associated base object.
