+++ SofaKernel/modules/SofaHelper/CMakeLists.txt
@@ -38,7 +38,6 @@ set(HEADER_FILES
     ${SRC_ROOT}/RandomGenerator.h
     ${SRC_ROOT}/SimpleTimer.h
     ${SRC_ROOT}/ScopedAdvancedTimer.h
-    ${SRC_ROOT}/SofaException.h
     ${SRC_ROOT}/SortedPermutation.h
     ${SRC_ROOT}/StringUtils.h
     ${SRC_ROOT}/TagFactory.h
@@ -127,7 +126,6 @@ set(SOURCE_FILES
     ${SRC_ROOT}/ScopedAdvancedTimer.cpp
     ${SRC_ROOT}/Polynomial_LD.cpp
     ${SRC_ROOT}/RandomGenerator.cpp
-    ${SRC_ROOT}/SofaException.cpp
     ${SRC_ROOT}/StringUtils.cpp
     ${SRC_ROOT}/TagFactory.cpp
     ${SRC_ROOT}/Utils.cpp
